Make block searching on dates useful by using a range of the whole day
