+++ packages/frontend-core/src/utils/lucene.js
@@ -99,8 +99,16 @@ export const buildLuceneQuery = filter => {
     filter.forEach(expression => {
       let { operator, field, type, value, externalType } = expression
       // Parse all values into correct types
-      if (type === "datetime" && value) {
-        value = new Date(value).toISOString()
+      if (type === "datetime") {
+        // Ensure date value is a valid date and parse into correct format
+        if (!value) {
+          return
+        }
+        try {
+          value = new Date(value).toISOString()
+        } catch (error) {
+          return
+        }
       }
       if (type === "number" && !Array.isArray(value)) {
         if (operator === "oneOf") {
